refactor: pick out utilsCipher to own file
