+++ src/app.js
@@ -1,7 +1,7 @@
 const errorHandler = require('./modules/customerError/errorHandler');
 const { validateOptionCLI } = require('./modules/validate/validateOptionCLI');
 const { validateConfigCipher } = require('./modules/validate/validateConfigCipher');
-const parseCLI = require('./modules/parseCLI/parseCLI');
+const { parseCLI } = require('./modules/parseCLI/parseCLI');
 const Streams = require('./modules/Streams/Streams');
 const isFileExists = require('./modules/validate/isFileExists');
