+++ src/modules/cipher/caesarCipher.js
@@ -1,12 +1,5 @@
 const { ENGLISH_ALFABIT } = require('../../const');
-
-const isUpperCase = (letter) => {
-  return letter === letter.toUpperCase();
-};
-
-const getPositiveShift = (shift) => {
-  return (shift % ENGLISH_ALFABIT.length) + ENGLISH_ALFABIT.length;
-};
+const { getPositiveShift, isUpperCase } = require('./utilsCipher');
 
 const getCipheredChar = (index, shift) => {
   return ENGLISH_ALFABIT[(index + getPositiveShift(shift)) % ENGLISH_ALFABIT.length];
@@ -24,4 +17,4 @@ const caesarCipher = (chunk, shift = 0) => {
   return cipheredChunk.join('');
 };
 
-module.exports = { isUpperCase, getPositiveShift, getCipheredChar, caesarCipher };
+module.exports = { caesarCipher, getCipheredChar };
