+++ src/modules/cipher/utilsCipher.js
@@ -1,0 +1,11 @@
+const { ENGLISH_ALFABIT } = require('../../const');
+
+const isUpperCase = (letter) => {
+  return letter === letter.toUpperCase();
+};
+
+const getPositiveShift = (shift) => {
+  return (shift % ENGLISH_ALFABIT.length) + ENGLISH_ALFABIT.length;
+};
+
+module.exports = { isUpperCase, getPositiveShift };
